% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whats_my_age_again.R
\name{whats_my_age_again}
\alias{whats_my_age_again}
\title{Calculate your age}
\usage{
whats_my_age_again(dob = NULL, enddate = NULL)
}
\arguments{
\item{dob}{A string of the form 'MM/DD/YYYY' indicated your birthday}

\item{enddate}{A string of the form 'MM/DD/YYYY' for another date to see how
old you were trsing}
}
\value{
lubridate object
}
\description{
Calculate your age in case you forget.
}
\examples{
\dontrun{
whats_my_age_again(dob = '02/09/1993', enddate = '02/09/2020'
}
}
